% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCombin_BLP.R
\name{regCombin_BLP}
\alias{regCombin_BLP}
\title{Bounds on the Best Linear Predictor (BLP) using Data Combination}
\usage{
regCombin_BLP(
  Ldata,
  Rdata,
  out_var,
  nc_var,
  c_var = NULL,
  w_var = NULL,
  w_x = NULL,
  w_y = NULL,
  nbCores = 1,
  DP = FALSE,
  full = FALSE,
  FW = FALSE,
  discrete = FALSE,
  pt_est = FALSE,
  bootstrap = FALSE,
  dK = 2.5,
  ASN = TRUE,
  unchanged = FALSE,
  factor = FALSE,
  K_sel = 0,
  dataset = 0,
  O_only = FALSE
)
}
\arguments{
\item{Ldata}{A data frame containing the dependent variable (`out_var`) and covariates.}

\item{Rdata}{A data frame containing the covariates used for regression (`nc_var`, `c_var`, `w_var`).}

\item{out_var}{Name of the dependent variable in `Ldata`.}

\item{nc_var}{Character vector of outside regressors (non-commonly observed) (used in regression).}

\item{c_var}{Character vector of covariates commonly observed (default NULL).}

\item{w_var}{Character vector of auxiliary covariates, available in both datasets but not entering the regression  (default NULL).}

\item{w_x}{Optional weights for `Rdata` observations (default NULL, equal weights used).}

\item{w_y}{Optional weights for `Ldata` observations (default NULL, equal weights used).}

\item{nbCores}{Number of CPU cores to use for parallel computations (default 1).}

\item{DP}{DEPRECATED, Logical, whether to also compute Pacini (2019) bounds}

\item{full}{Logical, full computation option (default FALSE).}

\item{FW}{Logical, whether to compute Frisch-Waugh type bounds (default FALSE).}

\item{discrete}{Logical, whether to treat conditioning variables as discrete (default FALSE).}

\item{pt_est}{Logical, whether to return point estimates only (default FALSE).}

\item{bootstrap}{Logical, whether to use bootstrap for variance estimation (default FALSE).}

\item{dK}{DEPRECATED, Numeric, parameter controlling smoothing or discretization (default 2.5).}

\item{ASN}{Logical, whether to compute asymptotic variance using influence function (default TRUE).}

\item{unchanged}{Logical, whether to keep all values when discretizing (default FALSE).}

\item{factor}{Logical, whether to treat W variables as factors (default FALSE).}

\item{K_sel}{Number of clusters if discrete clustering is used to partition the support of W (default 0, automatically selected).}

\item{dataset}{Integer flag indicating if we use propensity score weighting (default 0).}

\item{O_only}{Logical, whether to compute bounds only for outside regressors (default FALSE).}
}
\value{
A list containing:
\describe{
  \item{hull_point}{Matrix of BLP hull points for each covariate.}
  \item{mat_varb_out_unc}{Matrix of upper and lower bounds without constraints.}
  \item{mat_varb_out_unc_asn}{Matrix of asymptotic variance estimates.}
  \item{output_influence}{List of influence function calculations for asymptotic variance.}
}
}
\description{
Computes bounds for regression coefficients provided in D'Haultfoeuille, Gaillac, Maurel (2024) <doi:10.48550/arXiv.2412.04816>.
It optionally allows incorporating conditioning variables, weights, using bootstrap or subsampling instead of asymptotic normality,
and information about the support of the regressors.
}
\details{
The function proceeds in several steps:
1. Prepares matrices of covariates (`Xc_x`, `Xc_y`), outside regressors (`Xnc`), auxiliary variables (`Wc_x`, `Wc_y`).
2. Calls `compute_BLP()` to compute hull points (BLP bounds) based on the input datasets.
3. If `bootstrap` is TRUE, repeats computations via resampling to obtain variance estimates.
4. If `ASN` is TRUE, computes asymptotic variance using influence functions, optionally conditioning on `W`.
}
\examples{
\dontrun{
# Example usage
Ldata = data.frame(Y = rnorm(100), C1 = rnorm(100))
Rdata = data.frame(NC1 = rnorm(100), C1 = rnorm(100))

results <- regCombin_BLP(Ldata,Rdata,
                        out_var = "Y",
                        nc_var = c("NC1"),
                        c_var = c("C1"),
                        w_var=NULL,
                        w_x = NULL ,w_y= NULL,nbCores=1,
                        DP=FALSE,full=FALSE,FW=FALSE,discrete=TRUE,
                        pt_est=FALSE,bootstrap=FALSE,dK=2.5,
                        ASN=TRUE,unchanged =FALSE,factor=FALSE,K_sel= 0,
                        dataset=0, O_only =FALSE)

}

}
