% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_rule.R
\name{sampling_rule_blp_main}
\alias{sampling_rule_blp_main}
\title{The subsampling rule}
\usage{
sampling_rule_blp_main(n)
}
\arguments{
\item{n}{sample size.}
}
\value{
the subsampling size
}
\description{
The subsampling rule
}
